pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "google-generativeai>=0.8.4",
    "pytelegrambotapi>=4.26.0",
    "requests>=2.32.3",
]
